pyproject.toml
[project]
name = "hw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "transformers>=4.45.2",
    "pyright>=1.1.384",
    "pillow>=10.4.0",
    "jupyter>=1.1.1",
    "pip>=24.2",
    "scikit-learn>=1.5.2",
    "datasets>=3.1.0",
    "notebook>=7.2.2",
    "types-pillow>=10.2.0.20240822",
    "pytest>=8.3.3",
    "beartype>=0.19.0",
    "pyqt5-qt5>=5.15.2",
    "matplotlib>=3.9.2",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "wandb>=0.18.7",
    "jaxtyping>=0.2.36",
    "pytest-mpl>=0.17.0",
    "pytest-cases>=3.8.6",
    "pyqt5>=5.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
constraint-dependencies = ["pyqt5-qt5 <=5.15.2"]  # https://github.com/astral-sh/uv/issues/7005

[tool.ruff]
line-length = 120

[tool.pyright]
include = ["src"]
defineConstant = { DEBUG = true }
stubPath = "src/stubs"
pythonVersion = "3.12"
